# Conflicts:
#	backend/.idea/dataSources.xml
#	backend/src/main/resources/application.properties

+++ backend/src/main/java/org/example/demo/service/order_detail/OrderDetailService.java
@@ -3,21 +3,27 @@ package org.example.demo.service.order_detail;
 import jakarta.persistence.EntityManager;
 import lombok.extern.slf4j.Slf4j;
 import org.example.demo.dto.order.properties.request.OrderDetailRequestDTO;
+import org.example.demo.entity.order.core.Order;
+import org.example.demo.entity.order.enums.Type;
 import org.example.demo.entity.order.properties.OrderDetail;
 import org.example.demo.entity.product.core.ProductDetail;
 import org.example.demo.exception.CustomExceptions;
 import org.example.demo.repository.order.OrderProductDetailRepository;
+import org.example.demo.repository.order.OrderRepository;
 import org.example.demo.repository.order_detail.OrderDetailRepository;
 import org.example.demo.repository.product.core.ProductDetailRepository;
 import org.example.demo.service.IService;
 import org.example.demo.service.order.OrderService;
+import org.example.demo.util.event.EventUtil;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.data.domain.Page;
 import org.springframework.data.domain.Pageable;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
 
+import java.util.HashMap;
 import java.util.List;
+import java.util.Map;
 import java.util.Optional;
 
 @Slf4j
@@ -31,8 +37,12 @@ public class OrderDetailService implements IService<OrderDetail, Integer, OrderD
 
     @Autowired
     private OrderProductDetailRepository orderProductDetailRepository;
+
     @Autowired
     private ProductDetailRepository productDetailRepository;
+
+    @Autowired
+    private OrderRepository orderRepository;
 
     @Autowired
     private EntityManager entityManager;
@@ -57,28 +67,48 @@ public class OrderDetailService implements IService<OrderDetail, Integer, OrderD
     }
 
     @Override
+    @Transactional
     public OrderDetail save(OrderDetailRequestDTO requestDTO) {
-        System.out.println(requestDTO);
-        Optional<OrderDetail> entityFound = orderDetailRepository.findByOrderIdAndProductDetailId(requestDTO.getOrderId(), requestDTO.getProductDetailId());
+        // tìm kiếm hóa đơn chi tiết
+        Optional<OrderDetail> entityFound = orderDetailRepository.findByOrderIdAndProductDetailIdAndAverageDiscountEventPercent(requestDTO.getOrderId(), requestDTO.getProductDetailId(), requestDTO.getAverageDiscountEventPercent());
+        // tìm kiếm sản phẩm chi tiết
         Optional<ProductDetail> productDetailFound = productDetailRepository.findById(requestDTO.getProductDetailId());
-        System.out.println(entityFound.isPresent());
-        System.out.println(productDetailFound.isPresent());
+        // nếu tồn tại hóa đơn chi tiết và sản phẩm chi tiết
+        // ==> GHI ĐÈ HOẶC TẠO BẢN GHI MỚI
         if (entityFound.isPresent() && productDetailFound.isPresent()) {
+            entityFound.get().setDeleted(false);
+            ProductDetail productDetail = productDetailFound.get();
+            // cộng dồn số lượng cũ và mới
             int newCount = entityFound.get().getQuantity() + requestDTO.getQuantity();
+            // nếu không đủ số lượng đáp ứng
             if (newCount > productDetailFound.get().getQuantity()) {
                 throw new CustomExceptions.CustomBadRequest("Không đủ số lượng đáp ứng");
-            } else {
+            }
+            // ngược lại nếu đủ số lượng
+            else {
                 entityFound.get().setQuantity(newCount);
                 OrderDetail response = orderDetailRepository.save(entityFound.get());
                 orderService.reloadSubTotalOrder(response.getOrder());
                 return response;
             }
-        } else {
+        }
+        // ==> TẠO BẢN GHI MỚI
+        // nếu không tìm thấy hóa đơn chi tiết và sản phẩm chi tiết
+        else {
+            ProductDetail productDetail = productDetailRepository.findById(requestDTO.getProductDetailId()).orElseThrow(() -> new CustomExceptions.CustomBadRequest("Product detail not found"));
+            // tạo hóa đơn chi tiết mới
             OrderDetail orderDetail = new OrderDetail();
+            // set trạng thái xóa cho hóa đơn
             orderDetail.setDeleted(false);
+            // set số lượng
             orderDetail.setQuantity(requestDTO.getQuantity());
+            // set gia trị event trung bình
+            orderDetail.setAverageDiscountEventPercent(EventUtil.getAveragePercentEvent(productDetail.getProduct().getValidEvents()));
+            // set hóa đơn vào hóa đơn chi tiết
             orderDetail.setOrder(orderService.findById(requestDTO.getOrderId()));
-            orderDetail.setProductDetail(orderProductDetailRepository.findById(requestDTO.getProductDetailId()).orElseThrow(() -> new CustomExceptions.CustomBadRequest("Product detail not found")));
+            // set spct vào hóa đơn chi tiết
+            orderDetail.setProductDetail(productDetail);
+            // lưu lại hóa đơn chi tết
             OrderDetail response = orderDetailRepository.save(orderDetail);
             orderService.reloadSubTotalOrder(orderDetail.getOrder());
             return response;
@@ -90,22 +120,78 @@ public class OrderDetailService implements IService<OrderDetail, Integer, OrderD
         return null;
     }
 
+    @Transactional
     public OrderDetail updateQuantity(Integer integer, int newQuantity) {
+        // tìm hóa đơn chi tiết theo id
+        log.info("ĐANG TÌM HÓA ĐƠN");
         OrderDetail orderDetail = findById(integer);
+        log.info("ĐÃ TÌM THẤY HÓA ĐƠN");
+        // tìm hóa đơn
+        Order order = orderDetail.getOrder();
+        // số lương trong kho
         int quantityInStorage = orderDetail.getProductDetail().getQuantity();
+        // số lượng trong hóa đơn chi tiết
         int quantityInOrder = orderDetail.getQuantity();
 
+        // nếu số lượng mới lớn hơn trong kho
         if (newQuantity > quantityInStorage) {
             throw new CustomExceptions.CustomBadRequest("Không đủ số lượng đáp ứng");
-        } else if (newQuantity == 0) {
-            orderDetailRepository.delete(orderDetail);
+        }
+        //nếu số luọng về 0
+        else if (newQuantity == 0) {
+            // nếu đơn này đã thanh toán và là đơn online => thì chỉ xóa mềm
+            if (order.getIsPayment() && order.getType() == Type.ONLINE) {
+                orderDetail.setDeleted(true);
+                orderDetail.setQuantity(0);
+            }
+            // ngược lại xóa vĩnh viễn trưc tiếp
+            else {
+                orderDetailRepository.delete(orderDetail);
+                entityManager.flush();
+            }
+            // cập nhật lại giá trị
             orderService.reloadSubTotalOrder(orderDetail.getOrder());
             return orderDetail;
-        } else {
+        }
+        // nếu đủ số lượng đáp ứng
+        else {
+            orderDetail.setDeleted(false);
             orderDetail.setQuantity(newQuantity);
             orderService.reloadSubTotalOrder(orderDetail.getOrder());
             return orderDetailRepository.save(orderDetail);
         }
+    }
+
+    public Map<String, Boolean> checkAllowOverride(OrderDetailRequestDTO orderDetailRequestDTO) {
+        // lấy product detail
+        Optional<ProductDetail> productDetail = productDetailRepository.findById(orderDetailRequestDTO.getProductDetailId());
+        // lấy hóa đơn
+        Optional<Order> order = orderRepository.findById(orderDetailRequestDTO.getOrderId());
+        // lấy danh sách hóa đơm chi tiết dựa vào order id và product id
+        List<OrderDetail> orderDetailList = orderDetailRepository.findAllByOrderIdAndProductDetailId(
+                orderDetailRequestDTO.getOrderId(),
+                orderDetailRequestDTO.getProductDetailId()
+        );
+        // thu thập list % giảm giá
+        List<Double> listPercent = orderDetailList.stream().map(OrderDetail::getAverageDiscountEventPercent).toList();
+
+        Map<String, Boolean> map = new HashMap<String, Boolean>();
+        boolean changeOfEvent = false;
+        // nếu ko có hóa đơn chi tiết cũ làm có giảm giá
+        System.out.println(listPercent.toString());
+        if (!listPercent.isEmpty()) {
+            changeOfEvent = checkHasChangeOfEvent(productDetail.get(), listPercent);
+        }
+        log.info("HAS CHANGE EVENT: " + changeOfEvent);
+        map.put("hasChange", changeOfEvent);
+        return map;
+    }
+
+    public boolean checkHasChangeOfEvent(ProductDetail productDetail, List<Double> percentList) {
+        double newAverageDiscountEventPercent = EventUtil.getAveragePercentEvent(productDetail.getProduct().getValidEvents());
+        log.info("OLD : " + percentList.toString());
+        log.info("NEW : " + newAverageDiscountEventPercent);
+        return !percentList.contains(newAverageDiscountEventPercent);
     }
 
     public Page<OrderDetail> getPageOrderDetailByIdOrder(Integer id, Pageable pageable) {
